Fix all non-3rdparty compiler warnings across the project
Cast implicit int/float conversions, resolve sign-conversion and
unused-parameter/-variable warnings in core library headers, example
impl files, generator templates, and exploration scripts. Add missing
"signed": false to minireverb's count drop params to match their
size_t setters. Regenerate affected examples to propagate the
template fixes.

+++ JuceStandaloneGenerator/codegen_processor.py
@@ -133,7 +133,7 @@ def load_patches(blueprint: Blueprint) -> str:
                 result += f"""     if (auto* p = m_parameters.getParameter("{item['symbol']}"))
                         {{
                             const auto& range = m_parameters.getParameterRange("{item['symbol']}");
-                            float normalized = range.convertTo0to1(params.{item['symbol']});
+                            float normalized = range.convertTo0to1(static_cast<float>(params.{item['symbol']}));
                             p->setValueNotifyingHost(normalized);
                         }}
                 """
